fix: use bf16 in training arguments

+++ src/config/training.py
@@ -54,7 +54,7 @@ class TranslationTrainingConfig:
     lr_scheduler_type: Scheduler = Scheduler.linear
 
     # === Batching and gradients ===
-    fp16: bool = True
+    bf16: bool = True
     per_device_train_batch_size: int = 4
     per_device_eval_batch_size: int = 4
     gradient_accumulation_steps: int = 32
